% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IFBFat.R
\name{IFBFat}
\alias{IFBFat}
\title{IFBFat}
\usage{
IFBFat(dbpath, rumpfat, units, bcs, append = TRUE, summary = TRUE, sumcols)
}
\arguments{
\item{rumpfat}{name of column where rump fat measurement exists}

\item{units}{character. whether rump fat measurement is in "cm" or "mm"}

\item{bcs}{name of column where body condition score exists}

\item{append}{Logical. TRUE/FALSE. TRUE if IFBFat should be appended to original dataframe. Default is TRUE.}

\item{summary}{Logical. TRUE/FALSE. TRUE if a summary should be returned. Summary included mean, SE, and upper and lower 95\% CIs calculated by categories of choice.}

\item{sumcols}{vector of column names in database that should be aggregated on. Max categories == 3. Example: calculate summary of IFBFat by Sex and PMU}

\item{db}{dataframe of capture database}
}
\value{
Returns original capture database with IFBFat column appended
}
\description{
calculate IFBFat for bighorn sheep
}
\examples{
\donttest{fat<-IFBFat(dbpath = 'C:/Desktop/yourdb', rumpfat = 'MaxFat', units = "mm", bcs = 'BCS')}
}
\keyword{IFBFat,}
\keyword{condition,}
\keyword{nutritional}
\keyword{summary}
